bug: camp_id 컬럼 변경 수정
- entity, dto, serviceImpl에 명칭을 변경함

Fixes: #18

+++ lamdba-api-call/src/main/java/com/commit/lamdbaapicall/entity/CampingEntity.java
@@ -22,7 +22,7 @@ public class CampingEntity implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "camp_id")
-    private Integer campId;
+    private long campId;
 
     @Column(name = "camp_name", length = 100)
     private String campName;
